requirements.txt
httpx>=0.27,<1.0
pydantic>=2.6,<3.0
PyYAML>=6.0,<7.0
python-docx>=1.1,<2.0
google-generativeai>=0.8,<1.0
Jinja2>=3.1,<4.0
python-dateutil>=2.9,<3.0
